fix column bugs

+++ SMAA_LUT_MAIN/SMAA_LUT_MAIN/AreaMapGen.h
@@ -44,7 +44,7 @@ namespace LUT
 		void GenOrtho() noexcept;
 		void GenOrtho(int pattern, int left, int right, int offset_index);
 		size_t CalOrthoPixelPos(int pattern, int left, int right, int offset);
-		std::tuple<uint8_t, uint8_t> CalWight(float x1, float y1, float x2, float y2, float left);
+		std::tuple<float, float> CalWight(float x1, float y1, float x2, float y2, float left);
 	private:
 		DirectX::Image mAreaMap;
 	};
